+++ pallets/ddc-payouts/src/mock.rs
@@ -4,12 +4,13 @@
 
 use ddc_primitives::{
 	traits::{
-		cluster::{ClusterCreator, ClusterVisitor, ClusterVisitorError},
+		cluster::{ClusterCreator, ClusterProtocol},
 		customer::{CustomerCharger, CustomerDepositor},
 		pallet::PalletVisitor,
+		ClusterQuery,
 	},
-	ClusterBondingParams, ClusterFeesParams, ClusterGovParams, ClusterParams, ClusterPricingParams,
-	NodeType, DOLLARS,
+	ClusterBondingParams, ClusterFeesParams, ClusterParams, ClusterPricingParams,
+	ClusterProtocolParams, ClusterStatus, NodeType, DOLLARS,
 };
 use frame_election_provider_support::SortedListProvider;
 use frame_support::{
@@ -21,9 +22,11 @@ use frame_support::{
 use frame_system::mocking::{MockBlock, MockUncheckedExtrinsic};
 use sp_core::H256;
 use sp_io::TestExternalities;
+#[cfg(feature = "try-runtime")]
+use sp_runtime::TryRuntimeError;
 use sp_runtime::{
 	traits::{BlakeTwo256, Identity, IdentityLookup},
-	BuildStorage, DispatchError, Perquintill, TryRuntimeError,
+	BuildStorage, DispatchError, Perquintill,
 };
 use sp_std::prelude::*;
 
@@ -122,7 +125,7 @@ impl crate::pallet::Config for Test {
 	type Currency = Balances;
 	type CustomerCharger = TestCustomerCharger;
 	type CustomerDepositor = TestCustomerDepositor;
-	type ClusterVisitor = TestClusterVisitor;
+	type ClusterProtocol = TestClusterProtocol;
 	type TreasuryVisitor = TestTreasuryVisitor;
 	type NominatorsAndValidatorsList = TestValidatorVisitor<Self>;
 	type ClusterCreator = TestClusterCreator;
@@ -188,12 +191,12 @@ pub const ACCOUNT_ID_6: AccountId = 6;
 pub const ACCOUNT_ID_7: AccountId = 7;
 pub struct TestClusterCreator;
 impl<T: Config> ClusterCreator<T, Balance> for TestClusterCreator {
-	fn create_new_cluster(
+	fn create_cluster(
 		_cluster_id: ClusterId,
 		_cluster_manager_id: T::AccountId,
 		_cluster_reserve_id: T::AccountId,
 		_cluster_params: ClusterParams<T::AccountId>,
-		_cluster_gov_params: ClusterGovParams<Balance, BlockNumberFor<T>>,
+		_cluster_protocol_params: ClusterProtocolParams<Balance, BlockNumberFor<T>>,
 	) -> DispatchResult {
 		Ok(())
 	}
@@ -415,50 +418,81 @@ pub fn get_pricing(cluster_id: &ClusterId) -> ClusterPricingParams {
 	}
 }
 
-pub struct TestClusterVisitor;
-impl<T: Config> ClusterVisitor<T> for TestClusterVisitor {
-	fn ensure_cluster(_cluster_id: &ClusterId) -> Result<(), ClusterVisitorError> {
-		Ok(())
+pub struct TestClusterProtocol;
+impl<T: Config> ClusterQuery<T> for TestClusterProtocol {
+	fn cluster_exists(_cluster_id: &ClusterId) -> bool {
+		true
 	}
-	fn get_bond_size(
+
+	fn get_cluster_status(_cluster_id: &ClusterId) -> Result<ClusterStatus, DispatchError> {
+		unimplemented!()
+	}
+
+	fn get_manager_and_reserve_id(
 		_cluster_id: &ClusterId,
-		_node_type: NodeType,
-	) -> Result<u128, ClusterVisitorError> {
+	) -> Result<(T::AccountId, T::AccountId), DispatchError> {
+		unimplemented!()
+	}
+}
+
+impl<T: Config> ClusterProtocol<T, BalanceOf<T>> for TestClusterProtocol {
+	fn get_bond_size(_cluster_id: &ClusterId, _node_type: NodeType) -> Result<u128, DispatchError> {
 		Ok(10)
 	}
+
 	fn get_chill_delay(
 		_cluster_id: &ClusterId,
 		_node_type: NodeType,
-	) -> Result<BlockNumberFor<T>, ClusterVisitorError> {
+	) -> Result<BlockNumberFor<T>, DispatchError> {
 		Ok(BlockNumberFor::<T>::from(10u32))
 	}
+
 	fn get_unbonding_delay(
 		_cluster_id: &ClusterId,
 		_node_type: NodeType,
-	) -> Result<BlockNumberFor<T>, ClusterVisitorError> {
+	) -> Result<BlockNumberFor<T>, DispatchError> {
 		Ok(BlockNumberFor::<T>::from(10u32))
 	}
 
-	fn get_pricing_params(
-		cluster_id: &ClusterId,
-	) -> Result<ClusterPricingParams, ClusterVisitorError> {
+	fn get_pricing_params(cluster_id: &ClusterId) -> Result<ClusterPricingParams, DispatchError> {
 		Ok(get_pricing(cluster_id))
 	}
 
-	fn get_fees_params(cluster_id: &ClusterId) -> Result<ClusterFeesParams, ClusterVisitorError> {
+	fn get_fees_params(cluster_id: &ClusterId) -> Result<ClusterFeesParams, DispatchError> {
 		Ok(get_fees(cluster_id))
-	}
-
-	fn get_reserve_account_id(
-		_cluster_id: &ClusterId,
-	) -> Result<T::AccountId, ClusterVisitorError> {
-		let reserve_account = RESERVE_ACCOUNT_ID.to_ne_bytes();
-		Ok(T::AccountId::decode(&mut &reserve_account[..]).unwrap())
 	}
 
 	fn get_bonding_params(
 		_cluster_id: &ClusterId,
-	) -> Result<ClusterBondingParams<BlockNumberFor<T>>, ClusterVisitorError> {
+	) -> Result<ClusterBondingParams<BlockNumberFor<T>>, DispatchError> {
+		unimplemented!()
+	}
+
+	fn get_reserve_account_id(_cluster_id: &ClusterId) -> Result<T::AccountId, DispatchError> {
+		let reserve_account = RESERVE_ACCOUNT_ID.to_ne_bytes();
+		Ok(T::AccountId::decode(&mut &reserve_account[..]).unwrap())
+	}
+
+	fn activate_cluster_protocol(_cluster_id: &ClusterId) -> DispatchResult {
+		unimplemented!()
+	}
+
+	fn update_cluster_protocol(
+		_cluster_id: &ClusterId,
+		_cluster_protocol_params: ClusterProtocolParams<BalanceOf<T>, BlockNumberFor<T>>,
+	) -> DispatchResult {
+		unimplemented!()
+	}
+
+	fn bond_cluster(_cluster_id: &ClusterId) -> DispatchResult {
+		unimplemented!()
+	}
+
+	fn start_unbond_cluster(_cluster_id: &ClusterId) -> DispatchResult {
+		unimplemented!()
+	}
+
+	fn end_unbond_cluster(_cluster_id: &ClusterId) -> DispatchResult {
 		unimplemented!()
 	}
 }