+++ primitives/src/traits/treasury.rs
@@ -1,2 +1,0 @@
-pub trait TreasuryVisitor<T: Config> {
-    fn cluster_has_node(cluster_id: &ClusterId, node_pub_key: &NodePubKey) -> bool;